server/pyproject.toml
[project]
name = "visual-search-server"
version = "0.1.0"
description = "Stateless image segmentation + CLIP embedding server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "open-clip-torch>=2.26.1",
    "Pillow>=11.0.0",
    "numpy>=2.0.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "sam2>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]